Updates cost field to float

+++ models/item.go
@@ -3,5 +3,5 @@ package models
 type Item struct {
 	Name     string `json:"name"`
 	Desc     string `json:"description"`
-	GoldCost int    `json:"goldCost"`
+	GoldCost float64    `json:"goldCost"`
 }
